+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.1.0]
+
+### New features and improvements
+
+- Modal dialog
+    - SimpleModalBorder: add constructor to custom action type
+
+### Fixed bugs
+
+- Modal dialog
+    - SimpleModalBorder: fixed some code to customizable
+
 ## [1.0.0] - 2024-06-23
 
 - Initial release